1.0.18 | Import from SVN
This includes the following changelog:

= 1.0.18 =
- Fixed the Registration Plugin which had an error with registering new
users.

= 1.0.17r1 =
- Reupload the files without ending ?> in the classes to fix unexpected
output messages.

= 1.0.17 =
- Deprecated the Player class name from "Player" to "wpRPG_Player".
- Fixed a DebugFunction error that was in the Array causing it to not
preform as expected.

= 1.0.16 =
- Updated an issue in the Player Class which caused unexpected results
when updating a meta that had the value of 0, the function thought it
was a value of "false", and provided unexpected results.
- Added a Plugin_Class which will handle anything that's strictly
related to extended plugins such as the Plugin Manager.
- Added a plugin manager to visually see and control all plugins that
extend wpRPG. Plugins must register to use the manager.

= 1.0.15 =
- Created a Player Class which handles all Player_Meta functions
- Updated bundled plugins to use Player Class

= 1.0.14 =
- Fixed the Level Manager issue.
- Fixed a bunch of issues with the Debug functions.
- There was an issue causing the 1.0.3 Attack Module from working.

= 1.0.12 - 1.0.13 =
- Converted rpg_usermeta tables to native WP usermeta. Now instead of
custom sql and joins, you use get_user_meta and update_user_meta.
- Added a check for Cron Enable/Disabled tag. Wasn't properly in place
before.
- Removed the install for rpg_usermeta for reason listed above.
- Fixed the get_player_meta function for error checking.
- get_player_meta function also calls a new function 'get_meta' combines
all rpg_usermeta with wp_user and meta. Creates a super variable for the
player.
- Added a Reset Gold in the debug panel.
- Separated the Profile Pages from the Core class. Now the pages tab is
hooked by plugins to utilize.
- The Pages panel on the admin page also displays the shortcodes needed
for those pages. This is hooked with the plugins too.
- Completely removed races from the core class.

= 1.0.11 =
- Made extensive progress with the Profiles plugin. Took away the attack
profiles and used the Profile's native functionality. Also the members
plugin handles the list_players instead of the Attack plugin handling
it.
- Alot of changes have been made with the Attack plugin as well. Until
WP approves my WPRPG_Attack request, I've zipped and added the attack
plugin in the /plugins/ folder.
- You'll also have the opportunity to change the way the
Profiles/Members page looks and acts. Currently, it checks if your THEME
has a directory called "templates/wprpg/" and looks for the file
"view_profile.php" and "list_players.php". This allows you to customize
your Lists / Profiles how you see fit without worrying about my updates
screwing it up :-).
- Everything in my tests feels very stable so far. Using the new attack
module I included, you'll see that you can add things to the
List_Players actions and the Profile sections, like the attack button,
which is dynamically added rather than statically.
= 1.0.10 =
- Fixed a bug found in the plugins in which they weren't pointing to the
correct directories.

+++ wprpg-jscss.php
@@ -1,7 +1,10 @@
 <?php
 
 //check $_POST["content_txt"] is not empty
-if(isset($_POST["min_txt"]) && strlen($_POST["min_txt"])>0)
+/*if(!empty($_POST))
+	wp_die(var_dump($_POST));
+*/
+if(isset($_POST["min_txt"]) && strlen($_POST["min_txt"])>0 && isset($_POST["wprpg_levels"]))
 {
     //sanitize post value, PHP filter FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH
     $titleToSave = filter_var($_POST["title_txt"],FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH);
@@ -24,8 +27,44 @@ if(isset($_POST["min_txt"]) && strlen($_POST["min_txt"])>0)
         //output error
         header('HTTP/1.1 500 Looks like mysql error, could not insert record!');
         exit();
+    }    
+}
+elseif(isset($_POST["title_txt"]) && strlen($_POST["title_txt"])>0 && isset($_POST["wprpg_races"])){
+	//sanitize post value, PHP filter FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH
+    $titleToSave = filter_var($_POST["title_txt"],FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH);
+	if(isset($_POST['strength_txt']))
+		$strToSave = filter_var($_POST["strength_txt"],FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH);
+	else
+		$strToSave = 0;
+	if(isset($_POST['defense_txt']))
+		$defToSave = filter_var($_POST["defense_txt"],FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH);
+	else
+		$defToSave = 0;
+	if(isset($_POST['gold_txt']))
+		$goldToSave = filter_var($_POST["gold_txt"],FILTER_SANITIZE_STRING, FILTER_FLAG_STRIP_HIGH);
+	else
+		$goldToSave = 0;
+    //wp_die($contentToSave);
+	$wpdb->show_errors();
+    // Insert sanitize string in record
+    if($wpdb->insert($wpdb->prefix . 'rpg_races', array('strength'=>$strToSave, 'defense'=>$strToSave, 'gold'=>$goldToSave, 'title'=>$titleToSave), array('%d', '%d', '%d', '%s')))
+    {
+        //Record is successfully inserted, respond to ajax request
+        $my_id = $wpdb->insert_id; //Get ID of last inserted record from MySQL
+		echo '<tr id="item_'.$my_id.'"><td>'.$titleToSave.'</td>';
+		echo '<td>'.$strToSave.'</td>';
+		echo '<td>'.$defToSave.'</td>';
+		echo '<td>'.$goldToSave.'</td>';
+		echo '<td><div class="del_wrapper"><a href="#" class="del_button" id="del-'.$my_id.'">';
+		$path = plugins_url('images/icon_del.gif', __FILE__);
+		echo '<img src="'.$path .'" border="0" />';
+		echo '</a></div></td></tr>';   
+		$_POST = array();
+    }else{
+        //output error
+        header('HTTP/1.1 500 Looks like mysql error, could not insert record!');
+        exit();
     }
-    
 }
 elseif(isset($_POST["recordToDelete"]) && strlen($_POST["recordToDelete"])>0 && is_numeric($_POST["recordToDelete"]))
 {//do we have a delete request? $_POST["recordToDelete"]
@@ -33,10 +72,15 @@ elseif(isset($_POST["recordToDelete"]) && strlen($_POST["recordToDelete"])>0 && 
     //sanitize post value, PHP filter FILTER_SANITIZE_NUMBER_INT removes all characters except digits, plus and minus sign.
     $idToDelete = filter_var($_POST["recordToDelete"],FILTER_SANITIZE_NUMBER_INT);
     //try deleting record using the record ID we received from POST
-    if(!$wpdb->query("DELETE FROM ". $wpdb->prefix . "rpg_levels WHERE id=".$idToDelete))
-    {
-    }
-    
+	if(isset($_POST['wprpg_levels'])){
+		if(!$wpdb->query("DELETE FROM ". $wpdb->prefix . "rpg_levels WHERE id=".$idToDelete))
+		{
+		}
+    }elseif(isset($_POST['wprpg_races'])){
+		if(!$wpdb->query("DELETE FROM ". $wpdb->prefix . "rpg_races WHERE id=".$idToDelete))
+		{
+		}
+	}
 }
 
 
@@ -64,37 +108,71 @@ function wpRPG_get_plugin_code( ) {
 		"1"=>"//##### Add record when Add Record Button is clicked #########
 		$('#FormSubmit').click(function (e) {
 			e.preventDefault();
-			
-			if($('#title_txt').val()==='' || $('#min_txt').val()==='') //simple validation
-			{
-				alert('Please enter some text!');
-				return false;
+			if($('#wprpg_levels').length != 0){
+				if($('#title_txt').val()==='' || $('#min_txt').val()==='') //simple validation
+				{
+					alert('Please enter some text!');
+					return false;
+				}
+				
+				var myData = 'title_txt='+ $('#title_txt').val() +'&min_txt='+ $('#min_txt').val()+'&wprpg_levels='+ $('#wprpg_levels').val(); //post variables
+				
+				jQuery.ajax({
+					type: 'POST', // HTTP method POST or GET
+					url: '". site_url( 'wp-admin/admin-ajax.php' )."', //Where to make Ajax calls
+					dataType:'text', // Data type, HTML, json etc.
+					data:myData, //post variables
+					success:function(response){
+					$('#responds').append(response);
+					$('#title_txt').val('');
+					$('#min_txt').val('');//empty text field after successful submission
+					
+					},
+					error:function (xhr, ajaxOptions, thrownError){
+						alert(thrownError); //throw any errors
+					}
+				});
+			}
+			if($('#wprpg_races').length != 0){
+				if($('#title_txt').val()==='') //simple validation
+				{
+					alert('Please enter some text!');
+					return false;
+				}
+				
+				var myData = 'title_txt='+ $('#title_txt').val() +'&strength_txt='+ $('#strength_txt').val() +'&defense_txt='+ $('#defense_txt').val() +'&gold_txt='+ $('#gold_txt').val() + '&wprpg_races=1'; //post variables
+				
+				jQuery.ajax({
+					type: 'POST', // HTTP method POST or GET
+					url: '". site_url( 'wp-admin/admin-ajax.php' )."', //Where to make Ajax calls
+					dataType:'text', // Data type, HTML, json etc.
+					data:myData, //post variables
+					success:function(response){
+					$('#responds').append(response);
+					$('#title_txt').val('');
+					$('#strength_txt').val('0');
+					$('#gold_txt').val('0');
+					$('#defense_txt').val('0');//empty text field after successful submission
+					
+					},
+					error:function (xhr, ajaxOptions, thrownError){
+						alert(thrownError); //throw any errors
+					}
+				});
 			}
 			
-			var myData = 'title_txt='+ $('#title_txt').val() +'&min_txt='+ $('#min_txt').val(); //post variables
-			
-			jQuery.ajax({
-				type: 'POST', // HTTP method POST or GET
-				url: '". site_url( 'wp-admin/admin-ajax.php' )."', //Where to make Ajax calls
-				dataType:'text', // Data type, HTML, json etc.
-				data:myData, //post variables
-				success:function(response){
-				$('#responds').append(response);
-				$('#title_txt').val('');
-				$('#min_txt').val('');//empty text field after successful submission
-				
-				},
-				error:function (xhr, ajaxOptions, thrownError){
-					alert(thrownError); //throw any errors
-				}
-			});
 		}); " , "2"=>
 		"//##### Delete record when delete Button is clicked #########
 		$('a.del_button').click( function(e) {
 			e.preventDefault();
 			var clickedID = this.id.split('-'); //Split string (Split works as PHP explode)
 			var DbNumberID = clickedID[1]; //and get number from array
-			var myData = 'recordToDelete='+ DbNumberID; //build a post data structure
+			if($('#wprpg_races').length != 0){
+				var func = 'wprpg_races';
+			}else if($('#wprpg_levels').length != 0) {
+				var func = 'wprpg_levels';
+			}
+			var myData = 'recordToDelete='+ DbNumberID + '&' + func + '=1'; //build a post data structure
 			
 			jQuery.ajax({
 				type: 'POST', // HTTP method POST or GET
